implements EventGesturePan,Pinch,Twist - OS agnostic
this is shamelessy inspired by Federico-Ciuffardi's job on #36953

this is related to #13139 (and #37754 for 3.2 branch)

InputEventPan,Pinch,Twist are generated from InputEventScreenDrag,
thus a gesture made of n touches will generate n InputEventGestureXXX events.

InputEventGesturePan#get_delta, InputEventGesturePinch#get_factor and InputEventGestureTwist#get_rotation,
take this into account and are safe to be used as is.